Move models into Documents74 directory
