+++ tests/Doctrine/ODM/MongoDB/Tests/Functional/Ticket/GH2310Test.php
@@ -4,8 +4,8 @@ declare(strict_types=1);
 
 namespace Doctrine\ODM\MongoDB\Tests\Functional\Ticket;
 
-use Doctrine\ODM\MongoDB\Mapping\Annotations as ODM;
 use Doctrine\ODM\MongoDB\Tests\BaseTest;
+use Documents74\GH2310Container;
 
 class GH2310Test extends BaseTest
 {
@@ -48,28 +48,4 @@ class GH2310Test extends BaseTest
         self::assertSame($document->id, $result->id);
         self::assertNull($result->embedded);
     }
-}
-
-/**
- * @ODM\Document()
- */
-class GH2310Container
-{
-    /** @ODM\Id(strategy="none") */
-    public int $id;
-
-    /** @ODM\EmbedOne(targetDocument=GH2310Embedded::class, nullable=true) */
-    public ?GH2310Embedded $embedded;
-
-    public function __construct(int $id, ?GH2310Embedded $embedded)
-    {
-        $this->id = $id;
-        $this->embedded = $embedded;
-    }
-}
-
-class GH2310Embedded
-{
-    /** @ODM\Field(type="integer") */
-    public int $value;
 }
